NAS benchmark (stage 1) - move files (#5379)
