+++ nni/nas/benchmark/nasbench101/__init__.py
@@ -2,5 +2,5 @@
 # Licensed under the MIT license.
 
 from .constants import INPUT, OUTPUT, CONV3X3_BN_RELU, CONV1X1_BN_RELU, MAXPOOL3X3
-from .model import Nb101TrialStats, Nb101IntermediateStats, Nb101TrialConfig
+from .schema import Nb101TrialStats, Nb101IntermediateStats, Nb101TrialConfig
 from .query import query_nb101_trial_stats
