+++ nni/nas/benchmark/nasbench101/db_gen.py
@@ -6,8 +6,8 @@ import argparse
 from tqdm import tqdm
 from nasbench import api  # pylint: disable=import-error
 
-from nni.nas.benchmarks.utils import load_benchmark
-from .model import Nb101TrialConfig, Nb101TrialStats, Nb101IntermediateStats
+from nni.nas.benchmark.utils import load_benchmark
+from .schema import Nb101TrialConfig, Nb101TrialStats, Nb101IntermediateStats
 from .graph_util import nasbench_format_to_architecture_repr, hash_module
