+++ nni/nas/benchmark/nasbench201/__init__.py
@@ -2,5 +2,5 @@
 # Licensed under the MIT license.
 
 from .constants import NONE, SKIP_CONNECT, CONV_1X1, CONV_3X3, AVG_POOL_3X3
-from .model import Nb201TrialStats, Nb201IntermediateStats, Nb201TrialConfig
+from .schema import Nb201TrialStats, Nb201IntermediateStats, Nb201TrialConfig
 from .query import query_nb201_trial_stats
