+++ nni/nas/benchmark/nasbench201/query.py
@@ -6,26 +6,25 @@ import functools
 from peewee import fn
 from playhouse.shortcuts import model_to_dict
 
-from nni.nas.benchmarks.utils import load_benchmark
-from .model import Nb101TrialStats, Nb101TrialConfig, proxy
-from .graph_util import hash_module, infer_num_vertices
+from nni.nas.benchmark.utils import load_benchmark
+from .schema import Nb201TrialStats, Nb201TrialConfig, proxy
 
 
-def query_nb101_trial_stats(arch, num_epochs, isomorphism=True, reduction=None, include_intermediates=False):
+def query_nb201_trial_stats(arch, num_epochs, dataset, reduction=None, include_intermediates=False):
     """
-    Query trial stats of NAS-Bench-101 given conditions.
+    Query trial stats of NAS-Bench-201 given conditions.
 
     Parameters
     ----------
     arch : dict or None
         If a dict, it is in the format that is described in
-        :class:`nni.nas.benchmark.nasbench101.Nb101TrialConfig`. Only trial stats
+        :class:`nni.nas.benchmark.nasbench201.Nb201TrialConfig`. Only trial stats
         matched will be returned. If none, all architectures in the database will be matched.
     num_epochs : int or None
         If int, matching results will be returned. Otherwise a wildcard.
-    isomorphism : boolean
-        Whether to match essentially-same architecture, i.e., architecture with the
-        same graph-invariant hash value.
+    dataset : str or None
+        If specified, can be one of the dataset available in :class:`nni.nas.benchmark.nasbench201.Nb201TrialConfig`.
+        Otherwise a wildcard.
     reduction : str or None
         If 'none' or None, all trial stats will be returned directly.
         If 'mean', fields in trial stats will be averaged given the same trial config.
@@ -35,38 +34,36 @@ def query_nb101_trial_stats(arch, num_epochs, isomorphism=True, reduction=None, 
     Returns
     -------
     generator of dict
-        A generator of :class:`nni.nas.benchmark.nasbench101.Nb101TrialStats` objects,
+        A generator of :class:`nni.nas.benchmark.nasbench201.Nb201TrialStats` objects,
         where each of them has been converted into a dict.
     """
 
     if proxy.obj is None:
-        proxy.initialize(load_benchmark('nasbench101'))
+        proxy.initialize(load_benchmark('nasbench201'))
 
     fields = []
     if reduction == 'none':
         reduction = None
     if reduction == 'mean':
-        for field_name in Nb101TrialStats._meta.sorted_field_names:
-            if field_name not in ['id', 'config']:
-                fields.append(fn.AVG(getattr(Nb101TrialStats, field_name)).alias(field_name))
+        for field_name in Nb201TrialStats._meta.sorted_field_names:
+            if field_name not in ['id', 'config', 'seed']:
+                fields.append(fn.AVG(getattr(Nb201TrialStats, field_name)).alias(field_name))
     elif reduction is None:
-        fields.append(Nb101TrialStats)
+        fields.append(Nb201TrialStats)
     else:
         raise ValueError('Unsupported reduction: \'%s\'' % reduction)
-    query = Nb101TrialStats.select(*fields, Nb101TrialConfig).join(Nb101TrialConfig)
+    query = Nb201TrialStats.select(*fields, Nb201TrialConfig).join(Nb201TrialConfig)
     conditions = []
     if arch is not None:
-        if isomorphism:
-            num_vertices = infer_num_vertices(arch)
-            conditions.append(Nb101TrialConfig.hash == hash_module(arch, num_vertices))
-        else:
-            conditions.append(Nb101TrialConfig.arch == arch)
+        conditions.append(Nb201TrialConfig.arch == arch)
     if num_epochs is not None:
-        conditions.append(Nb101TrialConfig.num_epochs == num_epochs)
+        conditions.append(Nb201TrialConfig.num_epochs == num_epochs)
+    if dataset is not None:
+        conditions.append(Nb201TrialConfig.dataset == dataset)
     if conditions:
         query = query.where(functools.reduce(lambda a, b: a & b, conditions))
     if reduction is not None:
-        query = query.group_by(Nb101TrialStats.config)
+        query = query.group_by(Nb201TrialStats.config)
     for trial in query:
         if include_intermediates:
             data = model_to_dict(trial)
