+++ nni/nas/benchmark/nds/__init__.py
@@ -2,5 +2,5 @@
 # Licensed under the MIT license.
 
 from .constants import *
-from .model import NdsTrialConfig, NdsTrialStats, NdsIntermediateStats
+from .schema import NdsTrialConfig, NdsTrialStats, NdsIntermediateStats
 from .query import query_nds_trial_stats
