+++ nni/nas/benchmark/nlp/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from .model import NlpTrialStats, NlpIntermediateStats, NlpTrialConfig
+from .schema import NlpTrialStats, NlpIntermediateStats, NlpTrialConfig
 from .query import query_nlp_trial_stats
