--- nni/nas/benchmarks/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
-
-from .utils import load_benchmark, download_benchmark
